Merge in pull request by Chirag: Propose new pdf integration testing strategy #16
Test the PDF output using the UI with Selenium and the pages framework to set up the input data and the pdfbox library to assert against the output. This makes for tests that feel more natural to read and write and allows them to capture the true end-to-end experience of our system. This also helps to keep our mappers from diverging from the pages configuration.

As part of this change, some of the mappers were also updated to be null safe to allow for better test isolation in the pdf integration test, such that only the relevant pages in the UI need to be completed to test specific mappings. These mappers were previously not null safe because they implicitly baked in knowledge around page validations. In general, it seems like a better practice to design all of our mappers to be null safe such that there is no need to rely on these page configurations that could change.