Add serialVersionUID to every class that implements Serializable (#165)
During our last deploy, all of the users who were mid-application had their sessions invalidated, causing them to have to restart the app. This happened because we added fields to `UploadedDocument`, which inadvertently broke our session deserialization. 

**This commit attempts to rectify that, so that adding new fields will not break existing sessions. It also introduces a test which will fail if we break backwards compatibility**

Some background: Java expects you to provide a unique ID for any serializable class, called `serialVersionUID`. (This is one of the many things that are horrible about java's built-in serialization.) During serialization, the `serialVersionUID` is saved, and then during deserialization, Java uses this field to figure out the type of the resulting object. If you do not explicitly specify `serialVersionUID` in your serializable class, then Java will autogenerate one for you based on the instance fields and some other things.

We were not explicitly specifying `serialVersionUID`s for our serializable classes, so Java was autogenerating them for us. When we added fields to `UploadedDocument`, Java autogenerated a _different_ `serialVersionUID` for that class. Because the `serialVersionUID` was different, Java no longer knew what class it should use when deserializing existing sessions' uploaded docs.

This commit explicitly sets `serialVersionUID` for `UploadedDocument` and all other classes that implement `Serializable` to the exact values that were previously being autogenerated. As such, if we add fields to `UploadedDocument` in the future, Java will still understand that old sessions should be deserialized into the new, modified `UploadedDocument`.

To create the test for this PR, I ran the code from [our last release](https://github.com/codeforamerica/shiba/releases/tag/release-2021-04-20-1849), started a session, grabbed the serialized `ApplicationData` from the database for that session, and used it to create a new test fixture. The new test deserializes the old session's `ApplicationData` (thus verifying that we maintain backwards compatibility), and then uses it to generate CAF, CCAP, and UploadedDocument PDFs.

We may have to update this test fixture in the future and it may be annoying, so I tried to document its creation in a comment as well as I could.

## Notes
- We have a change in `main` that will cause all sessions to break on our next deploy. (It's because of this PR: https://github.com/codeforamerica/shiba/pull/159). To prevent that bad thing from happening, `Condition` needed to be moved back to the `inputconditions` package. Happy to explain this more if there are any skeptics
- This also introduces a TestUtils class
- The new test is fairly slow (~5 seconds locally), so I sped up `ApplicationDataEncryptorTest` to make up for it.

[finishes #177843919]

Co-authored-by: Luigi Ray-Montanez <luigi@codeforamerica.org>