Create journeyTest and pdfTest gradle tasks (#189)
This significantly speeds up test time in CI by creating  `journeyTest` and `pdfTest` tasks that can run in parallel with `unitTest` and `dbTest`. The tests in the new tasks were previously all in the `features` package

I renamed the abstract `FeatureTest` and the `features` package to `JourneyTest` and `journeys` to emphasize how expensive they are and highlight that all of them are in fact journey tests.