List of programs in success message is comma-separated
Before: You will receive a letter in the mail with next steps for your application for housing/emergency assistance/cash support in 7-10 days.

Now: You will receive a letter in the mail with next steps for your application for housing, emergency assistance and cash support in 7-10 days.

The Oxford comma is omitted because spanish does not use the Oxford comma.

[#178115548]